refactor:  - removed commented codes  - change to intuitive function names  - added more commented to explain logic and rationale
